also move pathfinder to geometry package
