+++ geometry/src/PathFinder_3rdParty.ts
@@ -1,5 +1,5 @@
-import { Heuristic, Pathfinder } from './Pathfinder';
-import { Coordinates } from '@duzh/geometry';
+import { Heuristic, Pathfinder } from './Pathfinder.js';
+import { Coordinates } from './Coordinates.js';
 import * as PF from 'pathfinding';
 import { DiagonalMovement } from 'pathfinding';
