+++ geometry/src/__tests__/CoordinatesUtils.test.ts
