+++ geometry/tsconfig.json
@@ -15,6 +15,6 @@
     "sourceMap": true,
     "strict": true,
     "plugins": [],
-    "types": ["node"]
+    "types": ["node", "vitest/config"]
   }
 }
