+++ main/package.json
@@ -14,7 +14,6 @@
   },
   "devDependencies": {
     "@types/node": "catalog:",
-    "@types/pathfinding": "catalog:pathfinding",
     "@vitest/coverage-v8": "catalog:vitest",
     "glob": "catalog:",
     "globals": "catalog:",
@@ -29,7 +28,6 @@
     "@duzh/geometry": "workspace:*",
     "@duzh/models": "workspace:*",
     "@duzh/utils": "workspace:*",
-    "pathfinding": "catalog:pathfinding",
     "zod": "catalog:zod"
   }
 }
