+++ main/src/main/maps/MapUtils.ts
@@ -4,9 +4,8 @@ import MapItem from '../objects/MapItem';
 import Door from '../objects/Door';
 import Block from '../objects/Block';
 import Bonus from '../objects/Bonus';
-import { Coordinates } from '@duzh/geometry';
+import { Coordinates, Heuristic, Pathfinder } from '@duzh/geometry';
 import { ObjectType } from '@main/objects/GameObject';
-import { Heuristic, Pathfinder } from '@main/geometry/Pathfinder';
 import Shrine from '@main/objects/Shrine';
 import Unit from '@main/units/Unit';
